Fixing tests

+++ tests/kltests.h
@@ -23,7 +23,7 @@ using namespace std;
 
 // Generate random messages
 extern const char *SAMPLE_MESSAGE;
-extern const UInt64 SAMPLE_MESSAGE_LEN;
+extern const size_t SAMPLE_MESSAGE_LEN;
 
 extern void makeRandomMessage(char *buffer, int msgId, int nRand);
 extern long long publishMessages(KLTopic *topic, int numMessages, int nRandMax, bool assertOffsets);
